+++ packages/chart/src/common/config.ts
@@ -1,0 +1,8 @@
+export const DEFAULT_VALS = {
+  MARGIN: {
+    top: 20,
+    right: 20,
+    bottom: 30,
+    left: 60,
+  },
+};